requirements.txt
pprint
firebase-admin
google-cloud-firestore
mock
attrs
boto3
